Remove "local dev mode"
This mode conflated many things. It's no longer needed as we now have
better mechanisms for launching configurations for local development in
both self-hosted and managed archodex.com configurations.

The one thing this does permanently change is removing the support for
using a local DynamoDB instance for a local managed archodex.com
development instance. But it's too hard to keep track of local and
remote account DBs, anyways.